Fix rest-doclet for endpoints with an array in RequestBody
REST document generation was failing for UserController.updateGenrePreferences(), which is mapped by a POST request to /users/{id}/library/genres
This is currently the only REST endpoint that requires an array in the request body.

The failure was occurring in generating the example for the call. It occurred because the parameter processing code removed the specialisation from the List<>, causing later code to generate an example list containing a bare Object, which cannot be serialised.

This fix preserves the specialisation, as already happens in the code that handles the endpoint response, so that later code works correctly.

!bug MUSAPI-314

!test
Run "gradle clean generateRestDocs" and check that the documentation is generated correctly.
If you are using the default setup, the documentation will be generated into: <api checkout directory>/build/rest-docs/index.html