Separate example object generator for request and response bodies
Instead of trying to instantiate an object of the correct type
create simple beans with the right set of properties.

For requests these are annotatated constructor parameters or
setter methods. (The @DocumentCommand annotation is no longer
needed on command objects.)

For responses these are getter methods.

The reason for doing this is that getter type methods on command
classes will no longer result in spurious fields being documented.
(See the tests section for examples.)

!BUG: MUSAPI-158
!TESTS:
Documentation should be largely unchanged but POST /promotions
should no longer include startDateBeforeEndDate and POST /groups
should no longer include id in the member example.