Handle parameterized types better
This is a major rewrite of the type handling mechanism in object
creation. However, it allows a map of maps to use the correct
types for the inner map.

Also allow controllers with no documentable methods. (They are just
ignored.)

!BUG: MUSAPI-158
!TESTS:
GET and PUT /tracks/{trackId}/licenses should now document the map
of maps that they return correctly.